Add files via upload
This is an SQL portfolio I worked on while training at Entrylevel where I had a role as a Data Analyst working for the charity, Education for All. I was asked by the Head of Fundraising to present the data on donor insights and donation rates.

Within the Fundraising team, my objectives were to:

- Increase the number of donors in my database
- Increase the donation frequency of my donors.
- Increase the value of donations in my database.

I had to present insights from the donation data to inform my fundraising strategy and increase donations.